requirements.txt
pymongo==4.3.3
discord.py==2.1.0
schedule==1.1.0
requests==2.28.1